requirements.txt
asgiref==3.6.0
attrs==22.2.0
Brotli==1.0.9
dj-database-url==1.2.0
Django==3.2.8
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.4.0
psycopg2-binary==2.9.5
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0.1
sqlparse==0.4.3
uritemplate==4.1.1
whitenoise==6.4.0
